Made Collection.__getitem__ not call Collection.list; fixed error where Collection.delete didn't pass on additional kwargs.
Calling Collection.list was a problem because it enumerates all the entities, which could be a large number. It also split the algorithm into two distinct, quite different cases. Now it calls the server directly with a particular name, and checks if the response contains multiple entries, so there will only ever be a small number of entities returned. The algorithm also collapses to a single case.