DVPL-3313: splunklib | Improve the SDK to make it easier to talk back to Splunk's API from within a modular input or custom search command
Part 2: splunklib.modularinput

Added Script.service property that returns a service object or None. Verified with automated unit tests, ad-hoc integration tests, and code review (thanks Shakeel).

The service object is created from the Server URI and Session Key passed to the command invocation on the input stream. It is available
as soon as the :code:`Script.stream_events` method is  called.   The script.service property returns a value of None, if you call it before the Script.stream_events` method is called.

Signed-off-by: David Noble <dnoble@splunk.com>