Fixed final test failures :)
We've adapted to the fact that unittest2.TestCase does not implement assertIsNotNone. We've left it to another day to resolve this issue: unittest2.TestCase does not implement skipTest. We've left a to-do in code. A bug is forthcoming.

This fix is verified on all three platforms, Python 2.6/2.7, and Splunk 6.2/6.3.

Next up: Push to develop and publish to github.