Add KV Store support to the Python SDK
This change adds KV Store support to the Python SDK, which includes
two main pieces:

1. A couple of small fixes in binding.py, specifically around allowing
us to make POST requests which are not encoded with application/x-www-form-urlencoded.

2. A set of classes in client.py to make it easy to access the KV Store
API, e.g. for accessing configuration, data, etc.

Finally, it also adds tests to verify these fixes, and an example to show
how to use it.