Decode binary response
The old Python 2 code concatenates the http-received data with a string,
which no longer works as Python 3 receives http data in binary form.

Simple decoding might be naive, as it might break non-utf8 or (or possibly
end up in the middle of a codepoint, not sure), but it works afaict.