Use __ when **query params are like owner/app/...
When one of the query parameters is, for instance, owner, app or sharing, as when sending a POST request to the + /acl for updating permissions, this call would normally fail because those are stripped out of the actual self.http.post() and used only for self._abspath(). This is a workaround for calling post() with those parameters starting with __. The query dict is then rebuilt stripping the __ so that in the end both self._abspath() and self.http.post() can receive owner, app and sharing arguments.