Add ToolRegistry (#4)
This change implements a ToolRegistry, containing a tool decorator, that is
used to register MCP tools.

The decorator infers a JSON schema from the input and output params of the decorated function.

Additionally to support Splunk-specific needs, this change introduces a ToolContext, which
tools can accept to gain access to additional functionalities. For now the ToolContext
exposes a connection to the Splunk REST API, but in future we will add more functionalities there
like: logging, tool cancellation, tool notifications and so on.